Move process.ts to /process folder
Make the process.ts steps more clearly organized in a separate folder
to match tfidf and extraction steps
